t3427: make it independent of the default branch name
At the moment, `git init` will use the name `master` for the initial
branch, but there are plans to use a more inclusive name in the future,
it seems the most likely candidate would be `main`.

Let's be ahead of the curve and adjust the test so that it will pass no
matter what might be the default branch name du jour.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>